Records of mean and standard deviation masses for species in BBS, from Dunning (2008)

From Dunning, raw records of mean and (where present) standard deviation for species in the BBS.

raw_masses

Format

A data frame with 895 rows and 21 variables:

species_id

also called aou, species code for BBS

id

species ID column

english_common_name

inherited

french_common_name

inherited

spanish_common_name

inherited

sporder

inherited

family

inherited

genus

Genus

species

Species

mass

mass (in g) from Dunning (2008)

sd

sd of mass from Dunning, if reported

sex

sex for record from Dunning, if reported

subspecies

subspecies from Dunning, if reported

location

location from Dunning, if reported

name_mismatch

1 or NA, whether the name in Dunning does not match the name in BBS. Typically slight spelling differences.

name_notes

name as listed in Dunning, if different from that listed in BBS

not_in_dunning

1 or NA, whether this species is not found in Dunning

close_genus

if a species is not in Dunning (2008) because of a name change or species split since 2008, genus of new classification

close_species

new species classification

close_subspecies

new subspecies classification

close_species_notes

description of species reclassification, if original as named is not in Dunning